specify json without indentation
also change wallet kind to Electrum with
uppercase E, and the alert_inputs so a service
could easily watch if the transaction is still
valid (without having to parse alert_tx)

Also, cancellation json does not need
to be aware of any details, except the
alert txid (without parsing the cancellation
tx), and timelock days - i.e. once the
alert txid was found, how urgent is it to
broadcast the cancellation.